Enviroment works again!
After some examination prepared in the previous commit, I realized that every time the enviroment exploded, it was on the verge of executing Eval(env,"(bind ?*state-p1* (create$))")
Anytime this expression (or a variant) happened the enviroment colapsed.
Thus, I decided to solve the need for using said expression INSIDE CLIPS, and not to rely on Eval for that.
Right now I do not know if the error was due to the create or the bind part.
There is so much code cleaning to be done...